refactor: made port and NIF implemenation share common code
