+++ rebar.config
@@ -8,3 +8,11 @@
   [{"(linux|darwin|solaris)", clean, "make -C c_src clean"},
    {"(freebsd)", clean, "gmake -C c_src clean"}]}.
 
+{profiles, [
+    {addr_san_test, [
+        {erl_opts, [
+                    debug_info
+                    , {d, 'TEST_ONLY_NIF', 1}
+                   ]}
+    ]}
+]}.
